+++ Magic/Spells/WaterSpell.cs
@@ -43,7 +43,7 @@ namespace Magic.Spells
 
                     dirt.state.Value = HoeDirt.watered;
 
-                    loc.temporarySprites.Add(new TemporaryAnimatedSprite(13, new Vector2(ix * (float)Game1.tileSize, iy * (float)Game1.tileSize), Color.White, 10, Game1.random.NextDouble() < 0.5, 70f, 0, Game1.tileSize, (float)(((double)iy * (double)Game1.tileSize + (double)(Game1.tileSize / 2)) / 10000.0 - 0.00999999977648258), -1, 0)
+                    loc.temporarySprites.Add(new TemporaryAnimatedSprite(13, new Vector2(ix * (float)Game1.tileSize, iy * (float)Game1.tileSize), Color.White, 10, Game1.random.NextDouble() < 0.5, 70f, 0, Game1.tileSize, (float)((iy * (double)Game1.tileSize + Game1.tileSize / 2) / 10000.0 - 0.00999999977648258), -1, 0)
                     {
                         delayBeforeAnimationStart = num * 10
                     });
